+++ apps/beets-frontend-v3/app/(app)/pools/[chain]/[variant]/[id]/swap/[[...txHash]]/page.tsx
@@ -1,63 +1,11 @@
 'use client'
 
-import { Box } from '@chakra-ui/react'
-import { PoolActionsLayout } from '@repo/lib/modules/pool/actions/PoolActionsLayout'
-import { getPoolActionableTokens } from '@repo/lib/modules/pool/pool-tokens.utils'
-import { usePoolRedirect } from '@repo/lib/modules/pool/pool.hooks'
-import { chainToSlugMap } from '@repo/lib/modules/pool/pool.utils'
-import { usePool } from '@repo/lib/modules/pool/PoolProvider'
-import { SwapForm } from '@repo/lib/modules/swap/SwapForm'
-import SwapLayout from '@repo/lib/modules/swap/SwapLayout'
-import { PathParams, SwapProviderProps } from '@repo/lib/modules/swap/SwapProvider'
-import { BalAlert } from '@repo/lib/shared/components/alerts/BalAlert'
-import { Hash } from 'viem'
+import { PoolSwapPage } from '@repo/lib/shared/pages/PoolSwapPage'
 
 type Props = {
   params: { txHash?: string[] }
 }
 // Page for swapping from a pool page
-export default function PoolSwapPage({ params: { txHash } }: Props) {
-  const { pool } = usePool()
-  const { redirectToPoolPage } = usePoolRedirect(pool)
-
-  const poolActionableTokens = getPoolActionableTokens(pool)
-
-  if (poolActionableTokens.length < 2) {
-    return (
-      <PoolActionsLayout>
-        <Box w="50%">
-          <BalAlert
-            content="You cannot swap the tokens in this pool because we are missing token metadata"
-            status="info"
-          />
-        </Box>
-      </PoolActionsLayout>
-    )
-  }
-
-  const maybeTxHash = (txHash?.[0] as Hash) || undefined
-
-  const pathParams: PathParams = {
-    chain: chainToSlugMap[pool.chain],
-    tokenIn: poolActionableTokens[0].address,
-    tokenOut: poolActionableTokens[1].address,
-    urlTxHash: maybeTxHash,
-  }
-  const props: SwapProviderProps = {
-    pathParams,
-    pool,
-    poolActionableTokens: poolActionableTokens,
-  }
-
-  const shouldRenderLayout = !!pool
-
-  return (
-    <PoolActionsLayout>
-      {shouldRenderLayout ? (
-        <SwapLayout props={props}>
-          <SwapForm redirectToPoolPage={redirectToPoolPage} />
-        </SwapLayout>
-      ) : null}
-    </PoolActionsLayout>
-  )
+export default function PoolSwapPageWrapper({ params: { txHash } }: Props) {
+  return <PoolSwapPage txHash={txHash} />
 }
